honeycomb_core::attributes

Trait AttributeBind

pub trait AttributeBind:
    Debug
    + Sized
    + Any {
    type StorageType: AttributeStorage<Self>;
    type IdentifierType: From<DartIdType> + ToPrimitive + Clone;

    const BIND_POLICY: OrbitPolicy;
}

Generic attribute trait

This trait is used to describe how a given attribute binds to the map, and how it should be stored in memory.

Example

A detailed example is provided in the user guide.

Required Associated Constants

const BIND_POLICY: OrbitPolicy

OrbitPolicy determining the kind of topological entity to which the attribute is associated.

Required Associated Types

type StorageType: AttributeStorage<Self>

Storage type used for the attribute.

type IdentifierType: From<DartIdType> + ToPrimitive + Clone

Identifier type of the entity the attribute is bound to.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.