honeycomb_core::attributes
Trait AttributeStorage

pub trait AttributeStorage<A: AttributeBind>: UnknownAttributeStorage {
    // Required methods
    fn set(&mut self, id: A::IdentifierType, val: A);
    fn get(&self, id: A::IdentifierType) -> Option<A>;
    fn replace(&mut self, id: A::IdentifierType, val: A) -> Option<A>;
    fn remove(&mut self, id: A::IdentifierType) -> Option<A>;

    // Provided method
    fn insert(&mut self, id: A::IdentifierType, val: A) { ... }
}

Common trait implemented by generic attribute storages.

This trait contain attribute-specific methods.

The documentation of this trait describe the behavior each function & method should have. “ID” and “index” are used interchangeably.

Required Methods

fn set(&mut self, id: A::IdentifierType, val: A)

Setter

Set the value of an element at a given index. This operation is not affected by the initial state of the edited entry.

Arguments

index: A::IdentifierType – Cell index.

val: A – Attribute value.

Panics

The method:

should panic if the index lands out of bounds

may panic if the index cannot be converted to usize

fn get(&self, id: A::IdentifierType) -> Option<A>

Getter

Arguments

index: A::IdentifierType – Cell index.

Return

The method should return:

Some(val: A) if there is an attribute associated with the specified index,

None if there is not.

Panics

The method:

should panic if the index lands out of bounds

may panic if the index cannot be converted to usize

fn replace(&mut self, id: A::IdentifierType, val: A) -> Option<A>

Setter

Replace the value of an element at a given index.

Arguments

index: A::IdentifierType – Cell index.

val: A – Attribute value.

Return

The method should return:

Some(val_old: A) if there was an attribute associated with the specified index,

None if there is not.

In both cases, the new value should be set to the one specified as argument.

Panics

The method:

should panic if the index lands out of bounds

may panic if the index cannot be converted to usize

fn remove(&mut self, id: A::IdentifierType) -> Option<A>

Remove an item from the storage and return it

Arguments

index: A::IdentifierType – Cell index.

Return

The method should return:

Some(val: A) if there was an attribute associated with the specified index,

None if there is not.

Panics

The method:

should panic if the index lands out of bounds

may panic if the index cannot be converted to usize

Provided Methods

fn insert(&mut self, id: A::IdentifierType, val: A)

Setter

Insert a value at a given empty index. Otherwise, see [#Panics] section for more information.

Arguments

index: A::IdentifierType – Cell index.

val: A – Attribute value.

Panics

The method:

should panic if there is already a value associated to the specified index

should panic if the index lands out of bounds

may panic if the index cannot be converted to usize