honeycomb_core::geometry

Struct Vector2

pub struct Vector2<T: CoordsFloat>(pub T, pub T);

2D vector structure

Generics

T: CoordsFloat – Generic FP type for coordinates.

Example

use honeycomb_core::geometry::Vector2;

let unit_x = Vector2::unit_x();
let unit_y = Vector2::unit_y();

assert_eq!(unit_x.dot(&unit_y), 0.0);
assert_eq!(unit_x.normal_dir()?, unit_y);

let two: f64 = 2.0;
let x_plus_y: Vector2<f64> = unit_x + unit_y;

assert_eq!(x_plus_y.norm(), two.sqrt());
assert_eq!(x_plus_y.unit_dir()?, Vector2(1.0 / two.sqrt(), 1.0 / two.sqrt()));

Tuple Fields
0: T1: T
Implementations

impl<T: CoordsFloat> Vector2<T>

pub fn unit_x() -> Self

Return a unit vector along the x axis.

pub fn unit_y() -> Self

Return a unit vector along the y axis.

pub fn into_inner(self) -> (T, T)

Consume self to return inner values.

pub fn x(&self) -> T

Return the value of the x coordinate of the vector.

pub fn y(&self) -> T

Return the value of the y coordinate of the vector.

pub fn norm(&self) -> T

Compute the norm of self.

pub fn unit_dir(&self) -> Result<Self, CoordsError>

Compute the direction of self as a unit vector.

Errors

This method will return an error if called on a null Vector2.

pub fn normal_dir(&self) -> Result<Vector2<T>, CoordsError>

Compute the direction of the normal vector to self as a unit vector.

Errors

This method will return an error if called on a null Vector2.

pub fn dot(&self, other: &Vector2<T>) -> T

Return the dot product between self and other.