honeycomb_core::geometry

Struct Vertex2

pub struct Vertex2<T: CoordsFloat>(pub T, pub T);

2D vertex structure

Attribute behavior

binds to 0-cells,

merge policy: the new vertex is placed at the midpoint between the two existing ones,

split policy: the current vertex is duplicated,

fallback policies: default implementations are used.

Generics

T: CoordsFloat – Generic FP type for coordinates.

Example

use honeycomb_core::geometry::{Vector2, Vertex2};

let v1 = Vertex2(1.0, 0.0);
let v2 = Vertex2(1.0, 1.0);

assert_eq!(v1.x(), 1.0);
assert_eq!(v1.y(), 0.0);

let two: f64 = 2.0;
// vectorAB = vertexB - vertexA
let v2_minus_v1: Vector2<f64> = v2 - v1;

assert_eq!(v2_minus_v1.norm(), 1.0);
assert_eq!(v2_minus_v1.unit_dir()?, Vector2::unit_y());

let mut v3 = Vertex2(0.0, 1.0);
// vertexA + vectorB = vertexA'
v3 += v2_minus_v1;

assert_eq!(v3.x(), 0.0);
assert_eq!(v3.y(), 2.0);

Tuple Fields
0: T1: T
Implementations

impl<T: CoordsFloat> Vertex2<T>

pub fn into_inner(self) -> (T, T)

Consume self to return inner values.

pub fn x(&self) -> T

Return the value of the x coordinate of the vertex.

pub fn y(&self) -> T

Return the value of the y coordinate of the vertex.

pub fn average(lhs: &Vertex2<T>, rhs: &Vertex2<T>) -> Vertex2<T>

Compute the mid-point between two vertices.

Panics

This function may panic if it cannot initialize an object T: CoordsFloat from the value 2.0. The chance of this happening when using T = f64 or T = f32 is most likely zero.

Example

use honeycomb_core::geometry::Vertex2;

let far_far_away: Vertex2<f64> = Vertex2(2.0, 2.0);
let origin: Vertex2<f64> = Vertex2::default();

assert_eq!(Vertex2::average(&origin, &far_far_away), Vertex2(1.0, 1.0));