Struct honeycomb_core::AttrSparseVec

pub struct AttrSparseVec<T: AttributeBind + AttributeUpdate> { /* private fields */ }

Custom storage structure for attributes

This structured is used to store user-defined attributes using a vector of Option<T> items. This means that valid attributes value may be separated by an arbitrary number of None.

This implementation should favor access logic over locality of reference.

Generics

T: AttributeBind + AttributeUpdate – Type of the stored attributes.

Example

Currently, this type is not meant to be used directly when operating on combinatorial maps, but it is kept public because it should eventually be part of the map building system where the user will add its own attributes and choose how they are stored. As such, no example is provided.

Implementations

impl<T: AttributeBind + AttributeUpdate + Clone> AttrSparseVec<T>

pub fn allocated_size(&self) -> usize

Available on crate feature utils only.

Return the amount of space allocated for the storage.

pub fn effective_size(&self) -> usize

Available on crate feature utils only.

Return the total amount of space used by the storage.

pub fn used_size(&self) -> usize

Available on crate feature utils only.

Return the amount of space used by valid entries of the storage.