Struct honeycomb_core::AttrStorageManager

pub struct AttrStorageManager { /* private fields */ }

Main attribute storage structure.

This structure is used to store all generic attributes that the user may add to the combinatorial map he’s building.

Implementation

The structure uses hashmaps in order to store each attribute’s dedicated storage. Which storage is used is determined by the associated type AttributeBind::StorageType.

The key type used by the map is each attribute’s TypeId. This implies that all attributes must have a different (unique) type, i.e. two decimal-valued attribute will need to be wrapped in respective dedicated structures.

Using the TypeId as the key value for collections yields a cleaner API, where the only argument passed to access methods is the ID of the cell of which they want the attribute. The actual attribute type is specified by passing a generic to the method. This bypasses any issues linked to literal-typed keys, such as typos, naming conventions, portability, etc.

Generics passed in access methods also have a secondary usage. To store heterogeneous collections, the internal hashmaps uses Box<dyn UnknownAttributeStorage> as their value type. Some cases require us to downcast the stored object (implementing UnknownAttributeStorage) to the correct collection type. This is achieved by using the downcast-rs crate and the associated storage type AttributeBind::StorageType. The code roughly looks like this:

pub struct Manager {
    inner: HashMap<TypeId, Box<dyn UnknownAttributeStorage>>,
}

impl Manager {
    pub fn add_storage<A: AttributeBind + 'static>(
        &mut self,
        size: usize,
    ) {
        let typeid = TypeId::of::<A>();
        let new_storage = <A as AttributeBind>::StorageType::new(size);
        self.inner.insert(typeid, Box::new(new_storage));
    }

    pub fn get_storage<A: AttributeBind>(&self) -> &<A as AttributeBind>::StorageType {
        let probably_storage = &self.inner[&TypeId::of::<A>()];
        // downcast is possible because:
        // - StorageType: AttributeStorage<A>
        // - AttributeStorage<A>: UnknownAttributeStorage
        probably_storage
            .downcast_ref::<<A as AttributeBind>::StorageType>()
            .expect("E: could not downcast generic storage to specified attribute type")
    }
}

Implementations

impl AttrStorageManager

pub fn extend_storages(&mut self, length: usize)

Extend the size of all storages in the manager.

Arguments

length: usize – Length by which storages should be extended.

pub fn merge_attributes( &mut self, orbit_policy: &OrbitPolicy<'_>, id_out: DartIdentifier, id_in_lhs: DartIdentifier, id_in_rhs: DartIdentifier, )

Execute a merging operation on all attributes associated with a given orbit for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn merge_vertex_attributes( &mut self, id_out: DartIdentifier, id_in_lhs: DartIdentifier, id_in_rhs: DartIdentifier, )

Execute a merging operation on all attributes associated with vertices for specified cells.

Arguments

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn merge_edge_attributes( &mut self, id_out: DartIdentifier, id_in_lhs: DartIdentifier, id_in_rhs: DartIdentifier, )

Execute a merging operation on all attributes associated with edges for specified cells.

Arguments

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn merge_face_attributes( &mut self, id_out: DartIdentifier, id_in_lhs: DartIdentifier, id_in_rhs: DartIdentifier, )

Execute a merging operation on all attributes associated with faces for specified cells.

Arguments

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn merge_other_attributes( &mut self, _orbit_policy: &OrbitPolicy<'_>, _id_out: DartIdentifier, _id_in_lhs: DartIdentifier, _id_in_rhs: DartIdentifier, )

Execute a merging operation on all attributes associated with a given orbit for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn split_attributes( &mut self, orbit_policy: &OrbitPolicy<'_>, id_out_lhs: DartIdentifier, id_out_rhs: DartIdentifier, id_in: DartIdentifier, )

Execute a splitting operation on all attributes associated with a given orbit for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.

pub fn split_vertex_attributes( &mut self, id_out_lhs: DartIdentifier, id_out_rhs: DartIdentifier, id_in: DartIdentifier, )

Execute a splitting operation on all attributes associated with vertices for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.

pub fn split_edge_attributes( &mut self, id_out_lhs: DartIdentifier, id_out_rhs: DartIdentifier, id_in: DartIdentifier, )

Execute a splitting operation on all attributes associated with edges for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.

pub fn split_face_attributes( &mut self, id_out_lhs: DartIdentifier, id_out_rhs: DartIdentifier, id_in: DartIdentifier, )

Execute a splitting operation on all attributes associated with faces for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.

pub fn split_other_attributes( &mut self, _orbit_policy: &OrbitPolicy<'_>, _id_out_lhs: DartIdentifier, _id_out_rhs: DartIdentifier, _id_in: DartIdentifier, )

Execute a splitting operation on all attributes associated with a given orbit for specified cells.

Arguments

orbit_policy: OrbitPolicy – Orbit associated with affected attributes.

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.

impl AttrStorageManager

pub fn add_storage<A: AttributeBind + 'static>( &mut self, size: usize, ) -> Result<(), ManagerError>

Add a new storage to the manager.

For a breakdown of the principles used for implementation, refer to the Explanation section of the AttrStorageManager documentation entry.

Arguments

size: usize – Initial size of the new storage.

Generic

A: AttributeBind + 'static – Type of the attribute that will be stored.

Return / Error

The function may return:

Ok(()) if the storage was successfully added,

Err(ManagerError::DuplicateStorage) if there was already a storage for the specified attribute.

pub fn extend_storage<A: AttributeBind>(&mut self, length: usize)

Extend the size of the storage of a given attribute.

Arguments

length: usize – Length by which the storage should be extended.

Generic

A: AttributeBind – Attribute of which the storage should be extended.

pub fn get_storage<A: AttributeBind>( &self, ) -> &<A as AttributeBind>::StorageType

Get a reference to the storage of a given attribute.

Generic

A: AttributeBind – Attribute stored by the fetched storage.

Panics

This method may panic if:

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

pub fn set_attribute<A: AttributeBind>(&mut self, id: A::IdentifierType, val: A)

Set the value of an attribute.

Arguments

id: A::IdentifierType – Cell ID to which the attribute is associated.

val: A – New value of the attribute for the given ID.

Generic

A: AttributeBind – Type of the attribute being set.

Panics

This method may panic if:

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

the index lands out of bounds

pub fn insert_attribute<A: AttributeBind>( &mut self, id: A::IdentifierType, val: A, )

Set the value of an attribute.

Arguments

id: A::IdentifierType – Cell ID to which the attribute is associated.

val: A – New value of the attribute for the given ID.

Generic

A: AttributeBind – Type of the attribute being set.

Panics

This method may panic if:

there already is a value associated to the given ID for the specified attribute

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

the index lands out of bounds

pub fn get_attribute<A: AttributeBind>( &self, id: A::IdentifierType, ) -> Option<A>

Get the value of an attribute.

Arguments

id: A::IdentifierType – Cell ID to which the attribute is associated.

Generic

A: AttributeBind – Type of the attribute fetched.

Return

The method may return:

Some(val: A) if there is an attribute associated with the specified index,

None if there is not.

Panics

This method may panic if:

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

the index lands out of bounds

pub fn replace_attribute<A: AttributeBind>( &mut self, id: A::IdentifierType, val: A, ) -> Option<A>

Set the value of an attribute.

Arguments

id: A::IdentifierType – ID of the cell to which the attribute is associated.

val: A – New value of the attribute for the given ID.

Generic

A: AttributeBind – Type of the attribute being set.

Return

The method should return:

Some(val_old: A) if there was an attribute associated with the specified index,

None if there was not.

Panics

This method may panic if:

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

the index lands out of bounds

pub fn remove_attribute<A: AttributeBind>( &mut self, id: A::IdentifierType, ) -> Option<A>

Remove the an item from an attribute storage.

Arguments

id: A::IdentifierType – Cell ID to which the attribute is associated.

Generic

A: AttributeBind – Type of the attribute fetched.

Return

The method may return:

Some(val: A) if was is an attribute associated with the specified index,

None if there was not.

Panics

This method may panic if:

there’s no storage associated with the specified attribute

downcasting Box<dyn UnknownAttributeStorage> to <A as AttributeBind>::StorageType fails

the index lands out of bounds

pub fn merge_attribute<A: AttributeBind>( &mut self, id_out: DartIdentifier, id_in_lhs: DartIdentifier, id_in_rhs: DartIdentifier, )

Merge given attribute values.

Arguments

id_out: DartIdentifier – Identifier to write the result to.

id_in_lhs: DartIdentifier – Identifier of one attribute value to merge.

id_in_rhs: DartIdentifier – Identifier of the other attribute value to merge.

pub fn split_attribute<A: AttributeBind>( &mut self, id_out_lhs: DartIdentifier, id_out_rhs: DartIdentifier, id_in: DartIdentifier, )

Split given attribute value.

Arguments

id_out_lhs: DartIdentifier – Identifier to write the result to.

id_out_rhs: DartIdentifier – Identifier to write the result to.

id_in: DartIdentifier – Identifier of the attribute value to split.