Trait honeycomb_core::AttributeBind

pub trait AttributeBind:
    Debug
    + Sized
    + Any {
    type StorageType: AttributeStorage<Self>;
    type IdentifierType: From<DartIdentifier> + ToPrimitive + Clone;

    // Required method
    fn binds_to<'a>() -> OrbitPolicy<'a>;
}

Generic attribute trait for support description

This trait can be implemented for a given attribute in order to hint at which components of the map the attribute is bound.

Example

Using the same context as the for the AttributeUpdate example, we can associate temperature to faces if we’re modeling a 2D mesh:

use honeycomb_core::{AttributeBind, AttributeUpdate, FaceIdentifier, OrbitPolicy, AttrSparseVec};

#[derive(Clone, Copy, Debug, PartialEq)]
pub struct Temperature {
    pub val: f32
}

impl AttributeBind for Temperature {
    type IdentifierType = FaceIdentifier;

    fn binds_to<'a>() -> OrbitPolicy<'a> {
        OrbitPolicy::Face
    }
}

Required Associated Types

type StorageType: AttributeStorage<Self>

Storage type used for the attribute.

type IdentifierType: From<DartIdentifier> + ToPrimitive + Clone

Identifier type of the entity the attribute is bound to.

Required Methods

fn binds_to<'a>() -> OrbitPolicy<'a>

Return an OrbitPolicy that can be used to identify the kind of topological entity to which the attribute is associated.

Object Safety

This trait is not object safe.