Trait honeycomb_core::AttributeUpdate

pub trait AttributeUpdate: Sized {
    // Required methods
    fn merge(attr1: Self, attr2: Self) -> Self;
    fn split(attr: Self) -> (Self, Self);

    // Provided method
    fn merge_undefined(attr: Option<Self>) -> Self { ... }
}

Generic attribute trait for logical behavior description

This trait can be implemented for a given attribute in order to define the behavior to follow when (un)sewing operations result in an update of the attribute.

Example

For an intensive property of a system (e.g. a temperature), an implementation would look like this:

use honeycomb_core::AttributeUpdate;

#[derive(Clone, Copy, Debug, PartialEq)]
pub struct Temperature {
    pub val: f32
}

impl AttributeUpdate for Temperature {
    fn merge(attr1: Self, attr2: Self) -> Self {
        Temperature { val: (attr1.val + attr2.val) / 2.0 }
    }

    fn split(attr: Self) -> (Self, Self) {
        (attr, attr)
    }

    fn merge_undefined(attr: Option<Self>) -> Self {
        attr.unwrap_or(Temperature { val: 0.0 })
    }
}

let t1 = Temperature { val: 273.0 };
let t2 = Temperature { val: 298.0 };

let t_new = AttributeUpdate::merge(t1, t2); // use AttributeUpdate::_
let t_ref = Temperature { val: 285.5 };

assert_eq!(Temperature::split(t_new), (t_ref, t_ref)); // or Temperature::_

Required Methods

fn merge(attr1: Self, attr2: Self) -> Self

Merging routine, i.e. how to obtain the new attribute value from the two existing ones.

fn split(attr: Self) -> (Self, Self)

Splitting routine, i.e. how to obtain the two attributes from a single one.

Provided Methods

fn merge_undefined(attr: Option<Self>) -> Self

Fallback merging routine, i.e. how to obtain the new attribute value from potentially undefined instances.

The default implementation may panic if no attribute can be used to create a value. The reason for that is as follows:

This trait and its methods were designed with the (un)sewing operation in mind. Their purpose is to simplify the code needed to propagate updates of attributes affected by the (un)sewing operation. Considering this context, as well as the definition of (un)linking operations, this panic seems reasonable: If the darts you are sewing have totally undefined attributes, you should most likely be linking them instead of sewing.

Object Safety

This trait is not object safe.