Crate honeycomb_core

honeycomb-core

This crate implements all basic structure and methods for 2D and 3D combinatorial map modeling.

This documentation focus on the implementation side of things and API usage, for more formal information about combinatorial maps, refer to the Definitions section of the user guide.

Features

Optional features can be enabled when compiling this crate:

utils – provides additionnal implementations for map generation, benchmarking & debugging

io – provides a function and a method for building maps from VTK meshes and vice versa

Structs

AttrCompactVec

Custom storage structure for attributes

AttrSparseVec

Custom storage structure for attributes

AttrStorageManager

Main attribute storage structure.

CMap2

Main map object.

CMapBuilder

Combinatorial map builder structure.

EdgeCollection

Edge ID collection

FaceCollection

Face ID collection

GridDescriptorutils

Grid description used to generate maps via CMapBuilder.

Orbit2

Generic 2D orbit implementation

Vector2

2D vector representation

Vertex2

2D vertex representation

VertexCollection

Vertex ID collection

Enums

BuilderError

Builder-level error enum

CMapError

Map-level error enum

CoordsError

Coordinates-level error enum

ManagerError

Attribute manager error enum.

OrbitPolicy

Orbit search policy enum.

Constants

NULL_DART_ID

Null value for dart identifiers

NULL_EDGE_ID

Null value for edge identifiers

NULL_FACE_ID

Null value for face identifiers

NULL_VERTEX_ID

Null value for vertex identifiers

NULL_VOLUME_ID

Null value for volume identifiers

Traits

AttributeBind

Generic attribute trait for support description

AttributeStorage

Common trait implemented by generic attribute storages.

AttributeUpdate

Generic attribute trait for logical behavior description

CoordsFloat

Common trait implemented by types used for coordinate representation.

UnknownAttributeStorage

Common trait implemented by generic attribute storages.

Type Aliases

DartIdentifier

Type definition for dart identifiers

EdgeIdentifier

Type definition for edge identifiers

FaceIdentifier

Type definition for face identifiers

VertexIdentifier

Type definition for vertex identifiers

VolumeIdentifier

Type definition for volume identifiers