Crate honeycomb_core

honeycomb-core

This crate implements all basic structure and methods for 2D and 3D combinatorial map modeling.

This documentation focus on the implementation side of things and API usage, for more formal information about combinatorial maps, refer to the Definitions section of the user guide.

Features

Optional features can be enabled when compiling this crate:

utils – provides additionnal implementations for map generation, benchmarking & debugging

io – provides a function and a method for building maps from VTK meshes and vice versa

Re-exports

pub use stm;

Modules

attributes

attribute modeling code

cmap

combinatorial maps implementations

geometry

geometry representation types & operators

prelude

commonly used items