Crate honeycomb_core

honeycomb-core

This crate implements all basic structures and methods for 2D and 3D combinatorial map modeling.

This documentation focus on the implementation side of things and API usage, for more formal information about combinatorial maps, refer to the Definitions section of the user guide.

Features

The par-internals feature can be enabled so that CMap structures use rayon internally to accelerate some methods (e.g. par_extend for new element additions). This may also lead to changes in performance due to first-touch mechanisms (to be confirmed).

Re-exports

pub use fast_stm as stm;

Modules

attributes

attribute modeling code

cmap

combinatorial maps implementations

geometry

geometry representation types & operators