Crate honeycomb_core

honeycomb-core

This crate implements all basic structures and methods for 2D and 3D combinatorial map modeling.

This documentation focus on the implementation side of things and API usage, for more formal information about combinatorial maps, refer to the Definitions section of the user guide.

Re-exports

pub use fast_stm as stm;

Modules

attributes

attribute modeling code

cmap

combinatorial maps implementations

geometry

geometry representation types & operators