honeycomb_core
Module geometry

geometry representation types & operators

This module contains all code related to custom spatial reprentation type implementations. This include custom vector / vertex types as well as generic traits for value encoding.

Structs

Vector2

2D vector representation

Vertex2

2D vertex representation

Enums

CoordsError

Coordinates-level error enum

Traits

CoordsFloat

Common trait implemented by types used for coordinate representation.